Fix dependency on a non-existent service "security.context"
The security part is optional in the role provider class, which is why `security.token_storage` is also defined with `on-invalid=ignore`.
As we have security disabled in dev env, neither security.token_storage nor security.context exists, but it still tries to replace the dependency and loses on-invalid=ignore. Then it was complaining, that the service is missing.