keep previous cookies
we recently fixed that we lose previous cookies. but the fix was keeping
the messages in the session and only setting the cookie on the last
redirect, which means the messages are lost if the session is destroyed
(e.g. redirect to page outside firewall)

changed to merge the flash messages cookie from requests into the
flashes. and using array_merge_recursive to keep multiple messages with
the same key.