Make fastly and cloudflare clients lazy loaded to support symfony secrets
author DemigodCode <info@bl-websolutions.de> 1702912036 +0100
committer David Buchmann <david.buchmann@liip.ch> 1702989567 +0100

If one wants to use secrets as parameters for fastly or cloudflare, the compiling of the container will fail.
After compiling the container the termination event will cause this error cause it's not necessary to have the decrypt key available when compiling the container and the service can't be initialized.

This fix creates the services as lazy services which will not make the options necessary on compile time of the container.